implement _get_lock, and do a bunch of refactoring.
